Rename types.ts to steps.ts
